Refactored sax parsing to cover styles and strings xml
